Factor out JPEG support into a module This is to make adding support for new formats easier. See the file pic_jpeg.h for the description of the interface that each image format must support.
Conflicts:

	Makefile
